requirements.txt
Django==5.0.1
djangorestframework==3.14.0
psycopg2-binary==2.9.11
python-decouple==3.8
pandas==2.2.0
openpyxl==3.1.2
rapidfuzz==3.6.1
unidecode==1.3.8
django-cors-headers==4.3.1
django-filter==23.5
gunicorn==23.0.0
dj-database-url==3.1.0
whitenoise==6.6.0
google-api-python-client==2.170.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
xlrd==2.0.1
playwright==1.51.0
celery[redis]==5.4.0
django-celery-beat==2.7.0
redis==5.0.8
openai==2.29.0
